% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_here.R
\name{from_here}
\alias{from_here}
\title{Get the relative path to a file}
\usage{
from_here(.path, .root = here::here())
}
\arguments{
\item{.path}{The path to the target file or directory.}

\item{.root}{The path to the parent directory. Defaults to project root (\code{here()}).}
}
\description{
\code{from_here()} returns the path to a file or directory using \code{here()}, relative
to a parent directory (by default, the project root).
}
\examples{
  from_here()
  \dontrun{
    myfilepath <- here("some", "path", "below", "your", "project", "root.txt")
    >> "parent/path/some/path/below/your/project/root.txt"
    myrelfilepath <- from_here(myfilepath)
    >> "some/path/below/your/project/root.txt"
    myrelfilepath <- from_here(myfilepath, here("some", "path")
    >> "below/your/project/root.txt"
  }
}
