% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_batch_geocoder.R
\name{here_batch_geocoder}
\alias{here_batch_geocoder}
\title{Batch Geocoding with Here API}
\usage{
here_batch_geocoder(df, search_var, country, here_id, here_code)
}
\arguments{
\item{df}{A data.frame containing the columns of the searchText and the Country where this address is located - data.frame/tibble}

\item{search_var}{The column which contains the address/text to be searched by the Here API}

\item{country}{The column which contains the country where this address is located}

\item{here_id}{The developer application ID that uniquely identify the user}

\item{here_code}{The developer application Code that uniquely identify the user}
}
\value{
Given a valid app_id, and app_code, the script returns a data.frame containing
the searched text/address, and the latitude and longitude, if it found any.
}
\description{
\code{here_batch_geocoder} Returns a data.frame containing the latitude and longitude of a list of addresses.
This should only be used to search large volumes of data with Here Geolocation Services.
}
\details{
This function is used by passing a data.frame, that contains a column with the text to be searched in the Here API,
and a column with the Country which this address/text refers to. It also leverages two other functions from this (heRe) package:
`here_get_job_status` and `here_download_job`. This process occurs to allow the user to upload a data.frame with many observations
to be fulfilled by latitude and longitude data from the Here API, and return a data.frame containing the searchText,
and the latitude and longitude pair.
}
\examples{

\dontrun{

df <-
  data.frame(address = c("27 King's College Cir, Toronto", "Avenida Joao Pessoa, Porto Alegre"),
             country = c("CAN", "BRA"))

here_batch_geocoder(df, search_var = address, country = country, '<YOUR_APP_ID>', '<YOUR_APP_CODE>')
}
}
