% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_geocoder.R
\name{here_geocoder}
\alias{here_geocoder}
\title{Geocoding with Here API}
\usage{
here_geocoder(src_address, here_id, here_code)
}
\arguments{
\item{src_address}{The address to be searched in the Here API}

\item{here_id}{The developer application ID that uniquely identify the user}

\item{here_code}{The developer application Code that uniquely identify the user}
}
\value{
Given a valid app_id, and app_code, the script returns a list containing
the latitude and longitude of the given address;
}
\description{
\code{here_geocoder} Returns the latitude and longitude of a given address.
}
\details{
This function works by verifying in the Here API the latitude and longitude
of a given address. The return is a list containing the latitude and longitude of the
searched address.
}
\examples{

\dontrun{
here_geocoder("27 King's College Cir, Toronto, ON M5S, Canada", '<YOUR_APP_ID>', '<YOUR_APP_CODE>')
}
}
