% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_get_job_status.R
\name{here_get_job_status}
\alias{here_get_job_status}
\title{Get Here Job Status by it's Id}
\usage{
here_get_job_status(job_id, here_id, here_code)
}
\arguments{
\item{job_id}{A string with the Job's unique Id.}

\item{here_id}{The developer application ID that uniquely identify the user}

\item{here_code}{The developer application Code that uniquely identify the user}
}
\value{
Given a valid job_id, here_id and here_code, it returns a string containing the Job's status.
}
\description{
\code{here_get_job_status} Returns the status of a job sent to run with the Batch Geocoding API (with `here_batch_geocoder`)
}
\details{
This is a support function to the `here_batch_geocoder`, to ensure that the job sent
has completed before trying to download it.
}
