% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_reverse_geocoder_landmarks.R
\name{here_reverse_geocoder_landmarks}
\alias{here_reverse_geocoder_landmarks}
\title{Reverse Geocoding with Here API - Gather Landmarks data}
\usage{
here_reverse_geocoder_landmarks(lat, lon, radius, here_id, here_code)
}
\arguments{
\item{lat}{The latitude to be searched using the Here API - numeric}

\item{lon}{The longitude to be searched using the Here API - numeric}

\item{radius}{The search radius in meters to be used by the Here API - numeric}

\item{here_id}{The developer application ID that uniquely identify the user}

\item{here_code}{The developer application Code that uniquely identify the user}
}
\value{
Given a valid app_id, and app_code, the script returns a data.frame containing
the closest landmarks, its distance to the given point, and each landmarks name, address,
country, state, city, district, street and postal code;
}
\description{
\code{here_reverse_geocoder_landmarks} Returns address information of the top 10 closest landmarks of a given latitude and longitude.
}
\details{
This function works by verifying in the Here API the top 10 closest landmarks of a given latitude and
longitude. The return is a data.frame containing multiple informations about the landmarks distance to the given point,
and each landmarks name, address, country, state, city, district, street and postal code.
}
\examples{

\dontrun{
here_reverse_geocoder_landmarks(43.643513, -79.378262, 5000, '<YOUR_APP_ID>', '<YOUR_APP_CODE>')
}
}
