% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/there.R
\name{there}
\alias{there}
\title{Unquoted file referencing}
\usage{
there(...)
}
\arguments{
\item{...}{\verb{[character]}\cr
Path components below the project root, can be empty.
Each argument should be a string containing one or more
path components separated by a forward slash \code{"/"}.}
}
\description{
\code{there()} is a wrapper for \code{\link[=here]{here()}} that allows you to reference files
relative to the project directory, without needing to quote each argument.
}
\examples{
\dontrun{
# Special characters are allowed, but they must be quoted.

## Do this:
there(my, project, file1.txt)
there("my/project/file1.txt")
there(`my/project/file1.txt`)
## Don't do this:
there(my/project/file1.txt)

## Do this:
there(my, project, "file-2.txt")
there(my, project, `file-2.txt`)
## Don't do this:
there(my, project, file-2.txt)
}
}
