% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_here.R
\name{dr_here}
\alias{dr_here}
\title{Situation report}
\usage{
dr_here(show_reason = TRUE)
}
\arguments{
\item{show_reason}{\verb{[logical(1)]}\cr
Include reason in output of \code{dr_here()}, defaults to \code{TRUE}.}
}
\description{
\code{dr_here()} shows a message that by default also includes the
reason why \code{here()} is set to a particular directory.
Use this function if \code{here()} gives unexpected results.
}
\examples{
dr_here()
}
