% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{here}
\alias{here}
\title{Find your files}
\usage{
here(...)
}
\arguments{
\item{...}{Path components below the project root, can be empty.}
}
\description{
Uses a reasonable heuristics to find your project's files, based on the
current working directory when the package is loaded.
}

