% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here-set.R
\name{set_here}
\alias{set_here}
\title{Add a marker file to a project}
\usage{
set_here(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{\verb{[character(1)]}\cr
Directory where to create \code{.here} file, defaults to the current directory.}

\item{verbose}{\verb{[logical(1)]}\cr
Verbose output, defaults to \code{TRUE}.}
}
\description{
\code{set_here()} creates an empty file named \code{.here}, by default
in the current directory.
When \code{here} encounters such a file, it uses the
directory that contains this file as root.
This is useful if none of the default criteria apply.
You need to restart the R session so that \code{here()}
picks up the newly created file.
}
